factor out resultsTable from relationTable
